Split configs into a separate file
